Namespace Trello classes
I have a requirement to create a new class, so I'm namespacing
`Squiddy::TrelloChecklist` into `Squiddy::Trello::Checklist`.

This moves it into a new directory and updates all the naming.
