+++ lib/squiddy/trello.rb
@@ -1,0 +1,6 @@
+require_relative 'trello/checklist'
+
+module Squiddy
+  class Trello
+  end
+end
